# cmd/crypto-vault-rest-api/docs/swagger.yaml
basePath: /api/v1/cvs
definitions:
  v1.BlobMetaResponse:
    properties:
      dateTimeCreated:
        description: Timestamp when the blob was created
        type: string
      encryptionKeyID:
        description: Optional encryption key ID for the blob
        type: string
      id:
        description: Unique identifier for the blob
        type: string
      name:
        description: Name of the blob
        type: string
      signKeyID:
        description: Optional signature key ID for the blob
        type: string
      signatureBlobID:
        type: string
      signatureFileName:
        type: string
      size:
        description: Size of the blob in bytes
        type: integer
      type:
        description: Type of the blob (e.g., file format)
        type: string
      userID:
        description: User who uploaded the blob
        type: string
    type: object
  v1.CryptoKeyMetaResponse:
    properties:
      algorithm:
        description: Cryptographic algorithm (e.g., AES, RSA, EC)
        type: string
      dateTimeCreated:
        description: Timestamp when the key was created
        type: string
      id:
        description: Unique identifier for the cryptographic key
        type: string
      keyPairID:
        description: Identifier for the key pair the key belongs to
        type: string
      keySize:
        description: Size of the cryptographic key
        type: integer
      type:
        description: Type of the cryptographic key (e.g., public, private)
        type: string
      userID:
        description: User who created the key
        type: string
    type: object
  v1.ErrorResponse:
    properties:
      message:
        description: The error message
        type: string
    type: object
  v1.InfoResponse:
    properties:
      message:
        description: The informational message
        type: string
    type: object
  v1.UploadKeyRequest:
    properties:
      algorithm:
        enum:
        - AES
        - RSA
        - ECDSA
        type: string
      key_size:
        type: integer
    type: object
info:
  contact:
    name: MGTheTrain
  description: Service capable of managing cryptographic keys and securing data at
    rest
  license:
    name: MIT license
    url: https://github.com/MGTheTrain/crypto-vault-service/blob/main/LICENSE
  title: CryptoVault Service API
  version: v1
paths:
  /blobs:
    get:
      consumes:
      - application/json
      description: Fetch a list of metadata for blobs based on query filters like
        name, size, type, and creation date.
      parameters:
      - description: Blob Name
        in: query
        name: name
        type: string
      - description: Blob Size
        in: query
        name: size
        type: integer
      - description: Blob Type
        in: query
        name: type
        type: string
      - description: Blob Creation Date (RFC3339)
        in: query
        name: dateTimeCreated
        type: string
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Offset the results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.BlobMetaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List blob metadata based on query parameters
      tags:
      - Blob
    post:
      consumes:
      - multipart/form-data
      description: Upload a blob to the system with optional encryption and signing
        using the provided keys
      parameters:
      - description: Blob File
        in: formData
        name: files
        required: true
        type: file
      - description: Encryption Key ID
        in: formData
        name: encryption_key_id
        type: string
      - description: Sign Key ID
        in: formData
        name: sign_key_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/v1.BlobMetaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Upload a blob with optional encryption and signing
      tags:
      - Blob
  /blobs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific blob and associated metadata by ID.
      parameters:
      - description: Blob ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/v1.InfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a blob by ID
      tags:
      - Blob
    get:
      consumes:
      - application/json
      description: Fetch the blob metadata by ID, including name, size, type, encryption
        and signing key IDs, and creation date.
      parameters:
      - description: Blob ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BlobMetaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Retrieve blob metadata by ID
      tags:
      - Blob
  /blobs/{id}/file:
    get:
      consumes:
      - application/json
      description: Download the content of a specific blob by ID, optionally decrypted
        with a provided decryption key ID.
      parameters:
      - description: Blob ID
        in: path
        name: id
        required: true
        type: string
      - description: Decryption Key ID
        in: query
        name: decryption_key_id
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Blob content
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Download a blob by ID
      tags:
      - Blob
  /blobs/{id}/signature:
    get:
      consumes:
      - application/json
      description: Download the signature file associated with a specific blob
      parameters:
      - description: Blob ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Signature file content
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Download a blob's signature by blob ID
      tags:
      - Blob
  /keys:
    get:
      consumes:
      - application/json
      description: Fetch a list of cryptographic key metadata based on filters like
        algorithm, type and creation date, with pagination and sorting options.
      parameters:
      - description: Cryptographic Algorithm
        in: query
        name: algorithm
        type: string
      - description: Key Type
        in: query
        name: type
        type: string
      - description: Key Creation Date (RFC3339)
        in: query
        name: dateTimeCreated
        type: string
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Offset the results
        in: query
        name: offset
        type: integer
      - description: Sort by a specific field
        in: query
        name: sortBy
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.CryptoKeyMetaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List cryptographic key metadata based on query parameters
      tags:
      - Key
    post:
      consumes:
      - application/json
      description: Generate cryptographic keys based on provided parameters and upload
        them to the system.
      parameters:
      - description: Cryptographic Key Data
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/v1.UploadKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/v1.CryptoKeyMetaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Upload cryptographic keys and metadata
      tags:
      - Key
  /keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific cryptographic key and associated metadata by
        ID.
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/v1.InfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a cryptographic key by ID
      tags:
      - Key
    get:
      consumes:
      - application/json
      description: Fetch the crypto key metadata by ID, including algorithm, key size
        and creation date.
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CryptoKeyMetaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Retrieve crypto key metadata by ID
      tags:
      - Key
  /keys/{id}/file:
    get:
      consumes:
      - application/json
      description: Download the content of a specific cryptographic key by ID in PEM
        format.
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/x-pem-file
      responses:
        "200":
          description: Cryptographic key content in PEM format
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Download a cryptographic key by ID
      tags:
      - Key
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"